docs: complete documentation updates for feature branch merge
- Created README.md, API_DOCUMENTATION.md, SETUP_GUIDE.md, MERGE_NOTES.md, TESTING_GUIDE.md

+++ Project Principle/demo/src/main/java/com/restaurant/demo/model/Manager.java
@@ -42,7 +42,7 @@ public class Manager extends Employee {
     public SalesReport viewSalesReport(List<CartItem> cartItems, List<User> users) {
         LocalDate today = LocalDate.now();
 
-        Map<Integer, List<CartItem>> todayCartByCustomer = cartItems.stream()
+        Map<Long, List<CartItem>> todayCartByCustomer = cartItems.stream()
                 .filter(item -> item.getAddedAt() != null && today.equals(item.getAddedAt().toLocalDate()))
                 .collect(Collectors.groupingBy(CartItem::getCustomerId));
